Refactor(sequencer)  improve code readability (#1247)
* improve code readability

* fix(sequencer):remove double import

* chore(sequencer): typo
